+++ src/SourceAdapters/LocalPathAdapter.php
@@ -3,8 +3,8 @@ declare(strict_types=1);
 
 namespace Plank\Mediable\SourceAdapters;
 
+use GuzzleHttp\Psr7\Utils;
 use Plank\Mediable\Helpers\File;
-use Plank\Mediable\Stream;
 use Psr\Http\Message\StreamInterface;
 
 /**
@@ -75,9 +75,9 @@ class LocalPathAdapter implements SourceAdapterInterface
     /**
      * {@inheritdoc}
      */
-    public function getStream(): ?StreamInterface
+    public function getStream(): StreamInterface
     {
-        return new Stream(fopen($this->path(), 'rb'));
+        return Utils::streamFor(fopen($this->path(), 'rb'));
     }
 
     /**