+++ src/SourceAdapters/RawContentAdapter.php
@@ -3,6 +3,7 @@ declare(strict_types=1);
 
 namespace Plank\Mediable\SourceAdapters;
 
+use GuzzleHttp\Psr7\Utils;
 use Plank\Mediable\Helpers\File;
 use Psr\Http\Message\StreamInterface;
 
@@ -74,9 +75,9 @@ class RawContentAdapter implements SourceAdapterInterface
     /**
      * {@inheritdoc}
      */
-    public function getStream(): ?StreamInterface
+    public function getStream(): StreamInterface
     {
-        return null;
+        return Utils::streamFor($this->source);
     }
 
     /**